refactor: [#30] MapperUtil 을 DishUtil 로 변경 및 내부 참조 객체 수정
