+++ backend/src/main/java/com/codesquad/sidedish/dish/DishService.java
@@ -5,7 +5,6 @@ import com.codesquad.sidedish.dish.dto.DishResponse;
 import com.codesquad.sidedish.dish.dto.Slice;
 import com.codesquad.sidedish.event_badge.EventBadgeRepository;
 import com.codesquad.sidedish.event_badge.domain.EventBadge;
-import com.codesquad.sidedish.util.MapperUtil;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +20,7 @@ public class DishService {
     public Slice<DishResponse> findDishBySection(String sectionName) {
         List<Dish> dishes = dishRepository.findBySectionName(sectionName);
         List<EventBadge> eventBadges = eventBadgeRepository.findAll();
-        MapperUtil.mapEventBadges(dishes, eventBadges);
+        DishUtil.mapEventBadges(dishes, eventBadges);
 
         List<DishResponse> responses = dishes.stream()
             .map(DishResponse::from)
@@ -32,7 +31,7 @@ public class DishService {
     public Slice<DishResponse> findDishByCategory(String categoryName) {
         List<Dish> dishes = dishRepository.findByCategoryName(categoryName);
         List<EventBadge> eventBadges = eventBadgeRepository.findAll();
-        MapperUtil.mapEventBadges(dishes, eventBadges);
+        DishUtil.mapEventBadges(dishes, eventBadges);
 
         List<DishResponse> responses = dishes.stream()
             .map(DishResponse::from)
